Fix return type for get_linearized_size.

+++ include/lbann/utils/opencv.hpp
@@ -108,8 +108,8 @@ inline cv::Mat get_opencv_mat(El::Matrix<uint8_t>& data, const std::vector<size_
 }
 
 /** Get the linearized size of dims. */
-inline int get_linearized_size(const std::vector<size_t>& dims) {
-  return std::accumulate(dims.begin(), dims.end(), 1, std::multiplies<int>());
+inline size_t get_linearized_size(const std::vector<size_t>& dims) {
+  return std::accumulate(dims.begin(), dims.end(), 1, std::multiplies<size_t>());
 }
 
 }  // namespace utils
